requirements.txt
gspread==6.2.1
google-auth==2.40.3
pillow==11.3.0
